requirements.txt
PyQt6
spotipy
python-dotenv
